ts-core-api/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.4.0; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
boltons==21.0.0
certifi==2021.10.8
charset-normalizer==2.0.9; python_version >= '3'
click==8.0.3; python_version >= '3.6'
databases==0.5.3
face==20.1.1
fastapi==0.70.1
glom==20.11.0
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.12.0; python_version >= '3.6'
httptools==0.3.0
idna==3.3; python_version >= '3'
jsonpickle==2.0.0
motor==2.5.1
pydantic==1.8.2
pymongo==3.12.3
python-dotenv==0.19.2
pyyaml==6.0
requests==2.26.0
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy==1.4.28; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
starlette==0.16.0; python_version >= '3.6'
typing-extensions==4.0.1; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn[standard]==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
